chore(digraph): rename component file name
